# Faster encoding