So I run a few benchmarks (also the one mentioned in #81) and flamegraphs and identified the vector resize as one problem when creating the pipelined commands.
With this small change I see some improvement.

redis-rs' own benchmarks:

```
name                                        master.txt ns/iter  faster-encoding.txt ns/iter    diff ns/iter  diff %
bench_simple_getsetdel                      37,027              37,401                                  374   1.01%
bench_simple_getsetdel_pipeline             17,230              16,305                                 -925  -5.37%
bench_simple_getsetdel_pipeline_precreated  16,370              15,409                                 -961  -5.87%
```

client-benchmark on master:

```
./target/release/client-benchmark  1.30s user 0.13s system 73% cpu 1.936 total
```

client-benchmark on faster-encoding:

```
./target/release/client-benchmark  0.91s user 0.12s system 66% cpu 1.538 total
```

hiredis does the same for the encoding (I just copied over the countdigits and bulklen functions and the totlen calculation)