@mitsuhiko 2016-06-12
That looks good. Going to pull this.